Move part to test in a component
